# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/comment": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "评论列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "content_identity",
                        "name": "content_identity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/add-collect": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "添加收藏",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "problem_identity",
                        "name": "problem_identity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/collect": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "收藏列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/collect-delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "取消收藏",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "problem_identity",
                        "name": "problem_identity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/comment-create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "评论创建",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "CommentCreate",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.CommentCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/comment-delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "评论删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user_identity",
                        "name": "user_identity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "comment_identity",
                        "name": "comment_identity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "define.CommentCreate": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "contentIdentity": {
                    "type": "string"
                }
            }
        },
...
    }
}